Fixed floating point register size issue
git-svn-id: https://subversion.assembla.com/svn/interp-mips@11 67dacdfc-b040-432d-96ee-e96a32a25a0d

+++ interp_mips/emulateMips.h
@@ -17,7 +17,7 @@ typedef struct
   uint32_t pc;
   int32_t gpr[32];
   uint32_t cpr0[32];
-  uint64_t cpr1[32];
+  uint32_t cpr1[32];
   uint32_t fcr1[5];
   int32_t lo;
   int32_t hi;
